requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
clarifai-grpc==8.0.0
google-api-core==2.7.2
google-auth==2.6.4
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
